Small formatting edits.

+++ tandem.sh
@@ -4,6 +4,7 @@ MIN_OVERLAP=0
 MAX_DISTANCE=100
 OUT_FILE_PREFIX=tandem
 K_NEAREST=2
+DO_CLEANUP=1
 
 min_overlap_set=0
 max_distance_set=0
@@ -20,6 +21,10 @@ fi
 for i in "$@"
 do
 case $i in
+    --no-cleanup)
+    DO_CLEANUP=0
+    shift
+    ;;
     --min-overlap=*)
     MIN_OVERLAP="${i#*=}"
     min_overlap_set=1
@@ -89,7 +94,8 @@ grep -F -f $1 $2 > $OUT_FILE_PREFIX.query.gff3
 
 
 
-bedtools closest -d -k $K_NEAREST -a <(grep $'\t'gene$'\t' $OUT_FILE_PREFIX.query.gff3 | sort -k1,1 -k4,4n) -b <(grep $'\t'gene$'\t' $2 | sort -k1,1 -k4,4n) | \
+# bedtools closest -d -k $K_NEAREST -a <(grep $'\t'gene$'\t' $OUT_FILE_PREFIX.query.gff3 | sort -k1,1 -k4,4n) -b <(grep $'\t'gene$'\t' $2 | sort -k1,1 -k4,4n) | \
+bedtools closest -d -k $K_NEAREST -a <(cat $OUT_FILE_PREFIX.query.gff3 | sort -k1,1 -k4,4g) -b <(cat $2 | sort -k1,1 -k4,4g) | \
   awk -v OFS='\t' '{ if ($19 == 0) {
                       if ($13 >= $4) {
                         $19 = -((($5<$14)?$5:$14) - $13 + 1);
@@ -111,14 +117,7 @@ bedtools closest -d -k $K_NEAREST -a <(grep $'\t'gene$'\t' $OUT_FILE_PREFIX.quer
 
 #Triticum_aestivum_CS42_TGACv1_scaffold_061102_1DL	82104	90815	-	80705	81710	+	TRIAE_CS42_1DL_TGACv1_061102_AA0185320	TRIAE_CS42_1DL_TGACv1_061102_AA0185310	394
 
-#awk -v OFS='\t' '{ if (($10 >= 0 || (($2 >= $5 && $3 >= $6) || ($5 >= $2 && $6 >= $3))) && ($4 == "+" && $7 == "+")) print $0 }' $OUT_FILE_PREFIX.tmp.tsv > $OUT_FILE_PREFIX.FF.tsv
-#awk -v OFS='\t' '{ if (($10 >= 0 || (($2 >= $5 && $3 >= $6) || ($5 >= $2 && $6 >= $3))) && ($4 == "-" && $7 == "-")) print $0 }' $OUT_FILE_PREFIX.tmp.tsv > $OUT_FILE_PREFIX.RR.tsv
-#awk -v OFS='\t' '{ if (($10 >= 0 || (($2 >= $5 && $3 >= $6) || ($5 >= $2 && $6 >= $3))) && ($4 == "+" && $7 == "-")) print $0 }' $OUT_FILE_PREFIX.tmp.tsv > $OUT_FILE_PREFIX.FR.tsv
-#awk -v OFS='\t' '{ if (($10 >= 0 || (($2 >= $5 && $3 >= $6) || ($5 >= $2 && $6 >= $3))) && ($4 == "-" && $7 == "+")) print $0 }' $OUT_FILE_PREFIX.tmp.tsv > $OUT_FILE_PREFIX.RF.tsv
-#awk -v OFS='\t' '{ if ($10 < 0 && (($2 >= $5 && $3 <= $6) || ($5 >= $2 && $6 <= $3))) print $0 }' $OUT_FILE_PREFIX.tmp.tsv > $OUT_FILE_PREFIX.full.tsv
-
-
-# 1     2       3       4 5       6       7 8                 9                 10
+# 1     2       3       4       5       6       7       8                       9                       10
 # Chr2	7410835	7415610	-	7409063	7409582	+	AT2G17050.TAIR10	AT2G17043.TAIR10	1253
 awk -v OFS='\t' '{ is_overlap=($10 < 0)?1:0; is_tandem=($3 < $5)?1:0; if ($4$7 == "++" && (is_overlap==1 || is_tandem==1)) print $0 }' $OUT_FILE_PREFIX.tmp.tsv | sort -u > $OUT_FILE_PREFIX.FF.tsv
 awk -v OFS='\t' '{ is_overlap=($10 < 0)?1:0; is_tandem=($3 < $5)?1:0; if ($4$7 == "--" && (is_overlap==1 || is_tandem==1)) print $0 }' $OUT_FILE_PREFIX.tmp.tsv | sort -u > $OUT_FILE_PREFIX.RR.tsv
@@ -127,5 +126,7 @@ awk -v OFS='\t' '{ is_overlap=($10 < 0)?1:0; is_tandem=($3 < $5)?1:0; if ($4$7 =
 awk -v OFS='\t' '{ is_overlap=($10 < 0)?1:0; if (is_overlap==1 && (($2 >= $5 && $3 <= $6) || ($5 >= $2 && $6 <= $3))) print $0; }' $OUT_FILE_PREFIX.tmp.tsv | sort -u > $OUT_FILE_PREFIX.full.tsv
 
 
-rm $OUT_FILE_PREFIX.query.gff3
-rm $OUT_FILE_PREFIX.tmp.tsv
+if [[ "$DO_CLEANUP" -eq 1 ]]; then
+  rm $OUT_FILE_PREFIX.query.gff3
+  rm $OUT_FILE_PREFIX.tmp.tsv
+fi
